feat(admin): add server-side session validation to dashboard layout (admin-auth.json auth-05)
- Convert dashboard layout.tsx from client to server component
- Extract client-side layout logic (sidebar, breadcrumbs) to dashboard-layout-client.tsx
- Import auth from @/lib/auth and call auth.api.getSession with request headers
- Redirect to /login if session is null/invalid
- Existing sidebar layout and breadcrumb rendering preserved in client component
