+++ apps/admin/src/app/(dashboard)/dashboard-layout-client.tsx
@@ -1,0 +1,77 @@
+"use client";
+
+import {
+  Breadcrumb,
+  BreadcrumbItem,
+  BreadcrumbLink,
+  BreadcrumbList,
+  BreadcrumbPage,
+  BreadcrumbSeparator,
+  Separator,
+  SidebarInset,
+  SidebarProvider,
+  SidebarTrigger,
+} from "@mpga/ui";
+import { usePathname } from "next/navigation";
+import { Fragment } from "react";
+
+import { AppSidebar } from "@/components/app-sidebar";
+
+function formatSegment(segment: string) {
+  return segment
+    .split("-")
+    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
+    .join(" ");
+}
+
+export function DashboardLayoutClient({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const pathname = usePathname();
+  const segments = pathname.split("/").filter(Boolean);
+
+  return (
+    <SidebarProvider>
+      <AppSidebar />
+      <SidebarInset>
+        <header className="flex h-16 shrink-0 items-center gap-2 px-4">
+          <SidebarTrigger className="-ml-1" />
+          <Separator
+            orientation="vertical"
+            className="mr-2 data-[orientation=vertical]:h-4"
+          />
+          <Breadcrumb>
+            <BreadcrumbList>
+              {segments.map((segment, index) => {
+                const href = "/" + segments.slice(0, index + 1).join("/");
+                const isLast = index === segments.length - 1;
+
+                return (
+                  <Fragment key={href}>
+                    {index > 0 && (
+                      <BreadcrumbSeparator className="hidden md:block" />
+                    )}
+                    <BreadcrumbItem>
+                      {isLast ? (
+                        <BreadcrumbPage>
+                          {formatSegment(segment)}
+                        </BreadcrumbPage>
+                      ) : (
+                        <BreadcrumbLink href={href} className="hidden md:block">
+                          {formatSegment(segment)}
+                        </BreadcrumbLink>
+                      )}
+                    </BreadcrumbItem>
+                  </Fragment>
+                );
+              })}
+            </BreadcrumbList>
+          </Breadcrumb>
+        </header>
+        <div className="flex-1 p-4">{children}</div>
+      </SidebarInset>
+    </SidebarProvider>
+  );
+}
