+++ apps/admin/src/app/(dashboard)/layout.tsx
@@ -1,77 +1,22 @@
-"use client";
+import { headers } from "next/headers";
+import { redirect } from "next/navigation";
 
-import {
-  Breadcrumb,
-  BreadcrumbItem,
-  BreadcrumbLink,
-  BreadcrumbList,
-  BreadcrumbPage,
-  BreadcrumbSeparator,
-  Separator,
-  SidebarInset,
-  SidebarProvider,
-  SidebarTrigger,
-} from "@mpga/ui";
-import { usePathname } from "next/navigation";
-import { Fragment } from "react";
+import { auth } from "@/lib/auth";
 
-import { AppSidebar } from "@/components/app-sidebar";
+import { DashboardLayoutClient } from "./dashboard-layout-client";
 
-function formatSegment(segment: string) {
-  return segment
-    .split("-")
-    .map((word) => word.charAt(0).toUpperCase() + word.slice(1))
-    .join(" ");
-}
-
-export default function DashboardLayout({
+export default async function DashboardLayout({
   children,
 }: {
   children: React.ReactNode;
 }) {
-  const pathname = usePathname();
-  const segments = pathname.split("/").filter(Boolean);
+  const session = await auth.api.getSession({
+    headers: await headers(),
+  });
 
-  return (
-    <SidebarProvider>
-      <AppSidebar />
-      <SidebarInset>
-        <header className="flex h-16 shrink-0 items-center gap-2 px-4">
-          <SidebarTrigger className="-ml-1" />
-          <Separator
-            orientation="vertical"
-            className="mr-2 data-[orientation=vertical]:h-4"
-          />
-          <Breadcrumb>
-            <BreadcrumbList>
-              {segments.map((segment, index) => {
-                const href = "/" + segments.slice(0, index + 1).join("/");
-                const isLast = index === segments.length - 1;
+  if (!session) {
+    redirect("/login");
+  }
 
-                return (
-                  <Fragment key={href}>
-                    {index > 0 && (
-                      <BreadcrumbSeparator className="hidden md:block" />
-                    )}
-                    <BreadcrumbItem>
-                      {isLast ? (
-                        <BreadcrumbPage>
-                          {formatSegment(segment)}
-                        </BreadcrumbPage>
-                      ) : (
-                        <BreadcrumbLink href={href} className="hidden md:block">
-                          {formatSegment(segment)}
-                        </BreadcrumbLink>
-                      )}
-                    </BreadcrumbItem>
-                  </Fragment>
-                );
-              })}
-            </BreadcrumbList>
-          </Breadcrumb>
-        </header>
-        <div className="flex-1 p-4">{children}</div>
-      </SidebarInset>
-    </SidebarProvider>
-  );
+  return <DashboardLayoutClient>{children}</DashboardLayoutClient>;
 }
